requirements.txt
asgiref==3.8.1
Django==5.1.1
django-phonenumber-field==8.0.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.15.2
phonenumbers==8.13.45
pillow==10.4.0
sqlparse==0.5.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
